Refactor PdfViewport component
Extract Overlay component. This refactor enables future changes to use
state around overlay items (like a state indicating what item is selected)
without re-rendering the entire viewport (especially the canvas)
